Add Components Directory and Components / MyPoint, Nav, Title
