+++ src/components/MyPoint.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+
+const MyTitle : React.FC = () => {
+    return(
+        <>
+        </>
+    );
+}
+
+export default MyTitle;
