+++ src/components/Title.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+
+const Title : React.FC = () => {
+    return(
+        <>
+        </>
+    );
+}
+
+export default Title;
