--- src/pages/Nav.tsx
@@ -1,14 +1,0 @@
-import React from "react";
-import NavStyle from "../styles/NavStyle";
-
-const Nav: React.FC = () => {
-    return(
-        <>
-            <NavStyle/>
-        
-        </>
-    );
-
-}
-
-export default Nav;
