+++ src/styles/NavStyle.tsx
@@ -1,10 +1,14 @@
 import styled from "styled-components";
 
 const NavStyle = styled.div`
+    display: flex;
+    flex-direction: row;
     width: 1920px;
-    height: 20px;
+    height: 40px;
     background-color: grey;
+
     
 `;
+
 
 export default NavStyle;
